+++ CHANGELOG.md
@@ -49,3 +49,6 @@ Now plugin always must receive `params` object
 * Update version of base bot
 
 ### 0.8.6
+
+* Fix attach like
+* Add refresh message marks